feature/refactored issues and categories modules

+++ backend/src/modules/issues/services/DeleteIssueService.ts
@@ -1,15 +1,15 @@
 import IIssueRepository from '@modules/issues/repositories/IIssuesRepository';
-// import AppError from '@shared/errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 class DeleteIssueService {
 	constructor(private issuesRepository: IIssueRepository) {}
 
 	public async execute(id: string): Promise<void> {
-		// const issue = await this.issuesRepository.findById(id);
+		const issue = await this.issuesRepository.findById(id);
 
-		// if (!issue) {
-		// 	throw new AppError('Issue not found! Error to delete');
-		// }
+		if (!issue) {
+			throw new AppError('Issue not found! Error to delete');
+		}
 
 		await this.issuesRepository.deleteIssue(id);
 	}
